Improvements of the windows samples:
- Renamed the Sample folder
- Proper usage of IDisposables in the Windows Samples
- Fixed namespace of the App class in the WPF sample
- Updated LibVlc.Windows to 3.0.4
