+++ Samples/Forms/LibVLCSharp.Forms.WPF.Sample/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="OpenTK" version="2.0.0" targetFramework="net47" />
   <package id="OpenTK.GLControl" version="1.1.2349.61993" targetFramework="net47" />
-  <package id="VideoLAN.LibVLC.Windows" version="3.0.0-alpha3" targetFramework="net461" />
+  <package id="VideoLAN.LibVLC.Windows" version="3.0.4" targetFramework="net47" />
   <package id="WpfLightToolkit" version="1.0.1" targetFramework="net47" />
   <package id="Xamarin.Forms" version="3.2.0.871581" targetFramework="net461" />
   <package id="Xamarin.Forms.Platform.WPF" version="3.2.0.871581" targetFramework="net461" />
