+++ Samples/LibVLCSharp.FSharp.Sample/LibVLCSharp.FSharp.Sample.fsproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="VideoLAN.LibVLC.Windows" Version="3.0.0-alpha3" />
+    <PackageReference Include="VideoLAN.LibVLC.Windows" Version="3.0.4" />
   </ItemGroup>
 
   <ItemGroup>
