+++ Samples/LibVLCSharp.Windows.Net40.Sample/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="VideoLAN.LibVLC.Windows" version="3.0.0" targetFramework="net40" />
+  <package id="VideoLAN.LibVLC.Windows" version="3.0.4" targetFramework="net40" />
 </packages>
