+++ Samples/LibVLCSharp.Windows.Sample/App.config
