+++ Samples/LibVLCSharp.Windows.Sample/LibVLCSharp.Windows.Sample.csproj
@@ -6,7 +6,7 @@
     <RuntimeIdentifiers>win7-x64;win7-x86</RuntimeIdentifiers>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="VideoLAN.LibVLC.Windows" Version="3.0.0-alpha3" />
+    <PackageReference Include="VideoLAN.LibVLC.Windows" Version="3.0.4" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\LibVLCSharp\LibVLCSharp.csproj" />
